generate schemas as JSON, without faking an endpoint with fumadocs
