+++ content/docs/Product-Opener/(api)/schemas/schemas/ingredient.mdx
@@ -1,18 +1,60 @@
 ---
 title: Ingredient Schema
+description: Schema definition for the Ingredient object in product data
 full: true
-_openapi:
-  method: GET
-  route: /ingredient
-  toc: []
-  structuredData:
-    headings: []
-    contents:
-      - content: Schema definition for Ingredient
 ---
 
-{/* This file was generated by Fumadocs. Do not edit this file directly. Any changes should be made by running the generation command again. */}
+# Ingredient Schema
 
-Schema definition for Ingredient
+This is a **data schema**, not an API endpoint. The `Ingredient` object describes the shape of
+data returned inside product records by the [Product Opener API](/docs/Product-Opener/api/).
 
-<APIPage document={"./specfiles-json/schemas-only.json"} operations={[{"path":"/ingredient","method":"get"}]} webhooks={[]} hasHead={false} />
+```json
+{
+  "type": "object",
+  "title": "ingredient",
+  "properties": {
+    "id": {
+      "type": "string"
+    },
+    "ingredients": {
+      "type": "array",
+      "description": "This structure gives the different ingredients and some information about them,\nlike estimate on their quantity.\n",
+      "items": {
+        "$ref": "#/components/schemas/Ingredient"
+      }
+    },
+    "percent": {
+      "type": "integer",
+      "description": "The percentage of the ingredient."
+    },
+    "percent_estimate": {
+      "type": "number",
+      "description": "Estimated percentage of the ingredient."
+    },
+    "percent_max": {
+      "type": [
+        "string",
+        "number"
+      ],
+      "description": "Maximum percentage of the ingredient."
+    },
+    "percent_min": {
+      "type": "integer",
+      "description": "Minimum percentage of the ingredient."
+    },
+    "text": {
+      "type": "string",
+      "description": "Text description of the ingredient."
+    },
+    "vegan": {
+      "type": "string",
+      "description": "Indicates if the ingredient is vegan."
+    },
+    "vegetarian": {
+      "type": "string",
+      "description": "Indicates if the ingredient is vegetarian."
+    }
+  }
+}
+```
