+++ content/docs/Product-Opener/(api)/schemas/schemas/product_attribute_groups.mdx
@@ -1,18 +1,81 @@
 ---
 title: Product-Attribute-Groups Schema
+description: Schema definition for the Product-Attribute-Groups object in product data
 full: true
-_openapi:
-  method: GET
-  route: /product_attribute_groups
-  toc: []
-  structuredData:
-    headings: []
-    contents:
-      - content: Schema definition for Product-Attribute-Groups
 ---
 
-{/* This file was generated by Fumadocs. Do not edit this file directly. Any changes should be made by running the generation command again. */}
+# Product-Attribute-Groups Schema
 
-Schema definition for Product-Attribute-Groups
+This is a **data schema**, not an API endpoint. The `Product-Attribute-Groups` object describes the shape of
+data returned inside product records by the [Product Opener API](/docs/Product-Opener/api/).
 
-<APIPage document={"./specfiles-json/schemas-only.json"} operations={[{"path":"/product_attribute_groups","method":"get"}]} webhooks={[]} hasHead={false} />
+```json
+{
+  "type": "object",
+  "description": "Specific data about a product to enable personal ranking\n",
+  "properties": {
+    "attribute_groups": {
+      "type": "array",
+      "description": "Each element is an attribute that can help compute a personal ranking for the product",
+      "items": {
+        "title": "product_attribute_group",
+        "type": "object",
+        "properties": {
+          "id": {
+            "type": "string",
+            "description": "Unique id of the attribute.\n\nIt will be use to match against preferences parameters.\n"
+          },
+          "status": {
+            "type": "string",
+            "enum": [
+              "known",
+              "unknown"
+            ],
+            "description": "wether we have the information to really compute this criteria or not."
+          },
+          "title": {
+            "type": "string",
+            "description": "A descriptive sentence about the situation of the product concerning attribute\n",
+            "example": "Does not contain: Molluscs"
+          },
+          "match": {
+            "type": "number",
+            "format": "float",
+            "minimum": 0,
+            "maximum": 100,
+            "description": "a numeric value for the match,\ntelling how much the products ranks well for this particular attribute.\nThe higher the value, the better the match.\n"
+          },
+          "grade": {
+            "description": "every attribute as a grade for a to e",
+            "type": "string",
+            "enum": [
+              "unknown",
+              "a",
+              "b",
+              "c",
+              "d",
+              "e"
+            ]
+          },
+          "name": {
+            "type": "string",
+            "description": "The name of attribute, for eventual display"
+          },
+          "icon_url": {
+            "type": "string",
+            "description": "an icon representing the attribute match (often using a color)"
+          },
+          "description": {
+            "type": "string",
+            "description": "An eventual description of the value of the property upon which this attribute is based"
+          },
+          "description_short": {
+            "type": "string",
+            "description": "An eventual short description of the value of the property upon which this attribute is based"
+          }
+        }
+      }
+    }
+  }
+}
+```
