+++ content/docs/Product-Opener/(api)/schemas/schemas/product_extended.mdx
@@ -1,18 +1,218 @@
 ---
 title: Product-Extended Schema
+description: Schema definition for the Product-Extended object in product data
 full: true
-_openapi:
-  method: GET
-  route: /product_extended
-  toc: []
-  structuredData:
-    headings: []
-    contents:
-      - content: Schema definition for Product-Extended
 ---
 
-{/* This file was generated by Fumadocs. Do not edit this file directly. Any changes should be made by running the generation command again. */}
+# Product-Extended Schema
 
-Schema definition for Product-Extended
+This is a **data schema**, not an API endpoint. The `Product-Extended` object describes the shape of
+data returned inside product records by the [Product Opener API](/docs/Product-Opener/api/).
 
-<APIPage document={"./specfiles-json/schemas-only.json"} operations={[{"path":"/product_extended","method":"get"}]} webhooks={[]} hasHead={false} />
+```json
+{
+  "title": "product_extended",
+  "type": "object",
+  "properties": {
+    "additives_original_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "additives_prev_original_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "added_countries_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "allergens_from_ingredients": {
+      "type": "string"
+    },
+    "allergens_from_user": {
+      "type": "string"
+    },
+    "amino_acids_prev_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "amino_acids_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "carbon_footprint_percent_of_known_ingredients": {
+      "type": "integer"
+    },
+    "categories_properties": {
+      "type": "object",
+      "properties": {
+        "agribalyse_food_code:en": {
+          "type": "string"
+        },
+        "agribalyse_proxy_food_code:en": {
+          "type": "string"
+        },
+        "ciqual_food_code:en": {
+          "type": "string"
+        }
+      }
+    },
+    "categories_properties_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "category_properties": {
+      "type": "object",
+      "additionalProperties": {
+        "description": "those are properties taken from the category taxonomy",
+        "type": "string"
+      }
+    },
+    "ciqual_food_name_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "compared_to_category": {
+      "type": "string",
+      "description": "the category to use for comparison.\n\n**TODO** explain how it is chosen.\n"
+    },
+    "conservation_conditions": {
+      "type": "string"
+    },
+    "customer_service": {
+      "type": "string",
+      "description": "Contact info of customer service.\n"
+    },
+    "expiration_date": {
+      "type": "string"
+    },
+    "link": {
+      "type": "string",
+      "description": "link to the product on the website of the producer\n"
+    },
+    "main_countries_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "minerals_prev_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "minerals_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "owner_fields": {
+      "type": "object",
+      "description": "Those are fields provided by the producer (through producers platform),\nand the value he provided.\n",
+      "properties": {
+        "additionalProperties": {
+          "description": "you can retrieve all kind of properties, the same as on the parent object (the product).\nIt's not processed entries (like tags for example) but raw ones.\n",
+          "oneOf": [
+            {
+              "type": "integer"
+            },
+            {
+              "type": "string"
+            },
+            {
+              "type": "object"
+            }
+          ]
+        }
+      }
+    },
+    "nova_groups_markers": {
+      "type": "object",
+      "description": "Detail of ingredients or processing that makes the products having Nova 3 or 4\n",
+      "propertyNames": {
+        "description": "Markers of level 3, 4, ...",
+        "type": "integer",
+        "enum": [
+          3,
+          4
+        ]
+      },
+      "additionalProperties": {
+        "type": "array",
+        "items": {
+          "type": "array",
+          "items": {
+            "description": "This array has two elements for each marker. The first element is the tag type of the marker (e.g. \"ingredients\", \"additives\", \"categories\"), and the second is the tag id in the taxonomy (e.g. \"en:e471\")",
+            "type": "string"
+          }
+        }
+      }
+    },
+    "nucleotides_tags": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    },
+    "origin": {
+      "type": "string"
+    },
+    "purchase_places": {
+      "type": "string",
+      "description": "Country, state, or city where the product can be purchased.\n",
+      "example": "Paris"
+    },
+    "purchase_places_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "stores": {
+      "type": "string",
+      "description": "Distributor name.\n",
+      "example": "Walmart"
+    },
+    "stores_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "traces_from_ingredients": {
+      "type": "string"
+    },
+    "traces_from_user": {
+      "type": "string"
+    }
+  },
+  "patternProperties": {
+    "conservation_conditions_(?<language_code>\\w\\w)": {
+      "type": "string"
+    },
+    "customer_service_(?<language_code>\\w\\w)": {
+      "type": "string"
+    },
+    "origin_(?<language_code>\\w\\w)": {
+      "type": "string",
+      "description": "`origin` in language indicated by `language_code`\n"
+    }
+  }
+}
+```
