+++ content/docs/Product-Opener/(api)/schemas/schemas/product_quality.mdx
@@ -1,18 +1,100 @@
 ---
 title: Product-Quality Schema
+description: Schema definition for the Product-Quality object in product data
 full: true
-_openapi:
-  method: GET
-  route: /product_quality
-  toc: []
-  structuredData:
-    headings: []
-    contents:
-      - content: Schema definition for Product-Quality
 ---
 
-{/* This file was generated by Fumadocs. Do not edit this file directly. Any changes should be made by running the generation command again. */}
+# Product-Quality Schema
 
-Schema definition for Product-Quality
+This is a **data schema**, not an API endpoint. The `Product-Quality` object describes the shape of
+data returned inside product records by the [Product Opener API](/docs/Product-Opener/api/).
 
-<APIPage document={"./specfiles-json/schemas-only.json"} operations={[{"path":"/product_quality","method":"get"}]} webhooks={[]} hasHead={false} />
+```json
+{
+  "type": "object",
+  "description": "This is data that is linked to products data quality\n",
+  "properties": {
+    "data_quality_bugs_tags": {
+      "type": "array",
+      "items": {
+        "type": "string",
+        "description": "Each data quality bug describes a data quality bug on the product.\n\nRefer to the [data quality taxonomy](https://static.openfoodfacts.org/data/taxonomies/data_quality.json)\n"
+      }
+    },
+    "data_quality_errors_tags": {
+      "type": "array",
+      "items": {
+        "type": "string",
+        "description": "Each Data Quality Error describes a data quality error on the product.\n\nRefer to the [data quality taxonomy](https://static.openfoodfacts.org/data/taxonomies/data_quality.json)\n"
+      }
+    },
+    "data_quality_info_tags": {
+      "type": "array",
+      "items": {
+        "type": "string",
+        "description": "Each Data Quality Info describes a data quality info on the product.\n\nRefer to the [data quality taxonomy](https://static.openfoodfacts.org/data/taxonomies/data_quality.json)\n"
+      }
+    },
+    "data_quality_tags": {
+      "type": "array",
+      "items": {
+        "type": "string",
+        "description": "Each Data Quality tag describes a data quality issue on the product.\n\nRefer to the [data quality taxonomy](https://static.openfoodfacts.org/data/taxonomies/data_quality.json)\n"
+      }
+    },
+    "data_quality_warnings_tags": {
+      "type": "array",
+      "items": {
+        "type": "string",
+        "description": "Each Data Quality warning tag describes a data quality warning on the product.\n\nRefer to the [data quality taxonomy](https://static.openfoodfacts.org/data/taxonomies/data_quality.json)\n"
+      }
+    },
+    "data_sources": {
+      "type": "string",
+      "description": "Source of data imported from producers.\n"
+    },
+    "data_sources_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "last_check_dates_tags": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "last_checked_t": {
+      "type": "integer"
+    },
+    "last_checker": {
+      "type": "string"
+    },
+    "states": {
+      "description": "comma separated list of values indicating some states of the product,\nlike things to be done, or to be completed.\nSee [states taxonomy](https://static.openfoodfacts.net/data/taxonomies/states.json)\n",
+      "type": "string"
+    },
+    "states_hierarchy": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    },
+    "states_tags": {
+      "type": "array",
+      "items": {
+        "description": "Each state describe something that is completed or is to be done or improved on the product.\n\nRefer to [states taxonomy](https://static.openfoodfacts.net/data/taxonomies/states.json)\n",
+        "type": "string"
+      }
+    },
+    "misc_tags": {
+      "description": "Information about different aspect of the product\n  Refer to [misc taxonomy](https://static.openfoodfacts.org/data/taxonomies/misc.json)\n",
+      "type": "array",
+      "items": {
+        "type": "string"
+      }
+    }
+  }
+}
+```
