+++ content/docs/Product-Opener/(docs)/dev/meta.json
@@ -1,0 +1,34 @@
+{
+  "title": "Introduction to the Product Opener documentation",
+  "pages": [
+    "index",
+    "how-to-quick-start-guide",
+    "how-to-develop-using-docker",
+    "how-to-develop-with-webcomponents",
+    "how-to-develop-producer-platform",
+    "how-to-use-gitpod",
+    "how-to-use-vscode",
+    "how-to-use-direnv",
+    "how-to-use-repl",
+    "how-to-write-and-run-tests",
+    "how-to-release",
+    "how-to-use-automated-pr-actions",
+    "how-to-use-pages-from-openfoodfacts-web",
+    "how-to-update-agribalyse-ecoscore",
+    "how-to-debug-minion-import-and-export-tasks",
+    "how-to-add-generic-recommendation-panel",
+    "how-to-learn-perl",
+    "explain-dynamic-test-grouping",
+    "explain-frontend-build-scripts",
+    "explain-javascript-usage-and-strategy",
+    "explain-nutrition-data",
+    "explain-packaging-data",
+    "explain-personal-search",
+    "explain-pro-dev-setup",
+    "explain-taxonomy-build-cache",
+    "explain-user-management",
+    "ref-docker-commands",
+    "ref-perl",
+    "ref-perl-pod"
+  ]
+}
