+++ scripts/generate-docs.mjs
@@ -53,23 +53,8 @@ rimrafSync(outNutripatrol, {
   },
 });
 
-// Run filter-schemas.js to create schemas-only.json
-execSync('node scripts/filter-schemas.mjs', { stdio: 'inherit' });
-
-// Generate schemas-only documentation first (API Schemas)
-// The schemas are in the /schemas/schemas folder, and we reference /schemas here so that the sidebar groups them and hides them until expanded
-void OpenAPI.generateFiles({
-  input: ['./specfiles-json/schemas-only.json'],
-  output: './content/docs/Product-Opener/(api)/schemas/schemas',
-  options: {
-    includeResponses: true,
-  },
-  includeDescription: true,
-  groupBy: false
-});
-
-// Run update-schemas.js to process the generated schema files to turn them to MDX with JSON code blocks
-execSync('node scripts/update-schemas.mjs', { stdio: 'inherit' });
+// Generate schema MDX files directly from the OpenAPI spec (no fake endpoints)
+execSync('node scripts/generate-schemas.mjs', { stdio: 'inherit' });
 
 // Generate v2 API documentation (API v2 Operations)
 void OpenAPI.generateFiles({
